requirements.txt
flask
google-generativeai
Pillow
opencv-python
numpy
matplotlib
